Liten forbedring av fakta-bosted-soker
