--- packages/fakta-bosted-soker/i18n/nb_NO.json
@@ -1,13 +1,0 @@
-{
-  "BostedSokerView.ForeignAddresse": "Utenlandsadresse",
-  "BostedSokerFaktaIndex.Soker": "Søker",
-
-  "Personstatus.Hjelpetekst": "Personstatus",
-  "Personstatus.Ukjent": "Ukjent",
-  "Sivilstand.Hjelpetekst": "Sivilstand",
-  "BostedSokerView.Region": "Region",
-
-  "BostedFaktaView.ForelderTo": "Forelder 2",
-  
-  "OppholdINorgeOgAdresserFaktaPanel.Parent": "Den andre forelderen"
-}
