+++ packages/fakta-bosted-soker/package.json
@@ -9,8 +9,7 @@
     "@fpsak-frontend/kodeverk": "1.0.0",
     "@fpsak-frontend/utils": "1.0.0",
     "@k9-sak-web/types": "1.0.0",
-    "react": "18.3.1",
-    "react-intl": "6.6.8"
+    "react": "18.3.1"
   },
   "devDependencies": {
     "@fpsak-frontend/utils-test": "1.0.0"
