+++ packages/fakta-bosted-soker/src/BostedSokerFaktaIndex.tsx
@@ -1,63 +1,27 @@
-import React from 'react';
-import { createIntl, createIntlCache, RawIntlProvider } from 'react-intl';
-
-import { KodeverkMedNavn } from '@k9-sak-web/types';
 import kodeverkTyper from '@fpsak-frontend/kodeverk/src/kodeverkTyper';
+import { KodeverkMedNavn } from '@k9-sak-web/types';
 
 import BostedSokerView from './components/BostedSokerView';
-import messages from '../i18n/nb_NO.json';
-
-const cache = createIntlCache();
-
-const intl = createIntl(
-  {
-    locale: 'nb-NO',
-    messages,
-  },
-  cache,
-);
-
-export type BostedSokerPersonopplysninger = {
-  navn: string;
-  adresser: [
-    {
-      adresseType: KodeverkMedNavn;
-      adresselinje1: string;
-      adresselinje2?: string;
-      adresselinje3?: string;
-      postNummer?: string;
-      poststed?: string;
-      land?: string;
-    },
-  ];
-  sivilstand: KodeverkMedNavn;
-  region: KodeverkMedNavn;
-  personstatus: KodeverkMedNavn;
-  avklartPersonstatus: {
-    overstyrtPersonstatus: KodeverkMedNavn;
-  };
-};
+import { BostedSokerPersonopplysninger } from './types';
 
 interface OwnProps {
   personopplysninger: BostedSokerPersonopplysninger;
-  sokerTypeTextId?: string;
+  sokerTypeText?: string;
   alleKodeverk: { [key: string]: KodeverkMedNavn[] };
 }
 
 const BostedSokerFaktaIndex = ({
   personopplysninger,
-  sokerTypeTextId = 'BostedSokerFaktaIndex.Soker',
+  sokerTypeText = 'Søker',
   alleKodeverk,
 }: OwnProps): JSX.Element => (
-  <RawIntlProvider value={intl}>
-    <BostedSokerView
-      personopplysninger={personopplysninger}
-      sokerTypeTextId={sokerTypeTextId}
-      regionTypes={alleKodeverk[kodeverkTyper.REGION]}
-      sivilstandTypes={alleKodeverk[kodeverkTyper.SIVILSTAND_TYPE]}
-      personstatusTypes={alleKodeverk[kodeverkTyper.PERSONSTATUS_TYPE]}
-    />
-  </RawIntlProvider>
+  <BostedSokerView
+    personopplysninger={personopplysninger}
+    sokerTypeText={sokerTypeText}
+    regionTypes={alleKodeverk[kodeverkTyper.REGION]}
+    sivilstandTypes={alleKodeverk[kodeverkTyper.SIVILSTAND_TYPE]}
+    personstatusTypes={alleKodeverk[kodeverkTyper.PERSONSTATUS_TYPE]}
+  />
 );
 
 export default BostedSokerFaktaIndex;
