--- packages/fakta-medlemskap/i18n/nb_NO.json
@@ -1,3 +1,0 @@
-{
-  "OppholdINorgeOgAdresserFaktaPanel.Parent": "Den andre forelderen"
-}
