+++ packages/fakta-medlemskap/src/components/oppholdInntektOgPerioder/OppholdINorgeOgAdresserFaktaPanel.tsx
@@ -97,7 +97,7 @@ const OppholdINorgeOgAdresserFaktaPanel: FunctionComponent<OppholdINorgeOgAdress
                 )}
                 {!f.isApplicant && (
                   <BostedSokerFaktaIndex
-                    sokerTypeTextId="OppholdINorgeOgAdresserFaktaPanel.Parent"
+                    sokerTypeText="Den andre forelderen"
                     personopplysninger={f.personopplysning}
                     alleKodeverk={alleKodeverk}
                   />
